complete initial move of main header bar
