+++ packages/app/package.json
@@ -84,6 +84,7 @@
         "graphql",
         "graphql-relay",
         "graphql/jsutils/Path",
+        "gravatar",
         "lodash",
         "mobx",
         "nanoid",
