+++ packages/frontend-shared/src/assets/logos/cypress-dark.png
