+++ packages/frontend-shared/src/assets/logos/cypress_s.png
