+++ packages/frontend-shared/src/gql-components/Auth.vue
