+++ packages/frontend-shared/src/gql-components/topnav/PromptContent.vue
