+++ packages/frontend-shared/src/gql-components/topnav/UserAvatar.vue
